+++ src/main/java/com/filmrental/entity/Payment.java
@@ -24,7 +24,7 @@ public class Payment {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name = "payment_id")
-	private Long paymentId;
+	private int paymentId;
 
 	@ManyToOne
 	@JoinColumn(name = "customer_id")
